+++ src/helpers/react-query/reactQueryErrorHandler.ts
@@ -1,0 +1,7 @@
+import { QueryKey } from "react-query";
+
+export const reactQueryErrorHandler = (key: QueryKey) => {
+  return (error: any) => {
+    console.log({ key, error: error.message });
+  };
+};
