+++ src/hooks/useBalance.ts
@@ -16,8 +16,8 @@ import {
 } from "src/constants/addresses";
 import { isTestnet } from "src/helpers";
 import { queryAssertion } from "src/helpers/react-query/queryAssertion";
+import { reactQueryErrorHandler } from "src/helpers/react-query/reactQueryErrorHandler";
 import { nonNullable } from "src/helpers/types/nonNullable";
-import { reactQueryErrorHandler } from "src/lib/react-query";
 
 import { useWeb3Context } from ".";
 import { useMultipleTokenContracts, useStaticFuseContract } from "./useContract";
