+++ src/hooks/useENS.ts
@@ -1,8 +1,8 @@
 import { useQuery } from "react-query";
 import { NetworkId } from "src/constants";
 import { queryAssertion } from "src/helpers/react-query/queryAssertion";
+import { reactQueryErrorHandler } from "src/helpers/react-query/reactQueryErrorHandler";
 import { nonNullable } from "src/helpers/types/nonNullable";
-import { reactQueryErrorHandler } from "src/lib/react-query";
 
 import { useWeb3Context } from ".";
