+++ src/hooks/useStakingRebaseRate.ts
@@ -4,7 +4,7 @@ import { SOHM_ADDRESSES, STAKING_ADDRESSES } from "src/constants/addresses";
 import { parseBigNumber } from "src/helpers";
 import { createDependentQuery } from "src/helpers/react-query/createDependentQuery";
 import { queryAssertion } from "src/helpers/react-query/queryAssertion";
-import { reactQueryErrorHandler } from "src/lib/react-query";
+import { reactQueryErrorHandler } from "src/helpers/react-query/reactQueryErrorHandler";
 
 import { useStaticSohmContract, useStaticStakingContract } from "./useContract";
