+++ src/views/Stake/components/StakeArea/components/RebaseTimer/hooks/useNextRebaseDate.ts
@@ -2,8 +2,8 @@ import { useQuery } from "react-query";
 import { NetworkId } from "src/constants";
 import { STAKING_ADDRESSES } from "src/constants/addresses";
 import { parseBigNumber } from "src/helpers";
+import { reactQueryErrorHandler } from "src/helpers/react-query/reactQueryErrorHandler";
 import { useStaticStakingContract } from "src/hooks/useContract";
-import { reactQueryErrorHandler } from "src/lib/react-query";
 
 export const nextRebaseDateQueryKey = () => ["useNextRebaseDate"];
